Add separate heartbeat e2e test
